convert float to decimal

+++ traderboard/async_tasks.py
@@ -1,5 +1,6 @@
 from celery import shared_task
 from Market import Market
+from decimal import Decimal
 from asgiref.sync import sync_to_async
 from .models import TradingAccount, AccountTrades, AccountTransactions
 
@@ -12,7 +13,7 @@ def add(x, y):
 def record_transaction(event, ta):
     if TradingAccount.objects.get(id=ta.id):
         asset = event['a']
-        amount = float(event['d'])
+        amount = Decimal(event['d'])
         side = 'DEPOSIT' if amount >= 0 else 'WITHDRAWAL'
         date = Market.to_datetime(event['E'])
         trans = AccountTransactions(account=ta,
@@ -34,8 +35,8 @@ def record_trade(event, ta):
     if TradingAccount.objects.get(id=ta.id):
         symbol = event['s']
         side = event['S']
-        quantity = float(event['q'])
-        price = float(event['p'])
+        quantity = Decimal(event['q'])
+        price = Decimal(event['p'])
         date = Market.to_datetime(event['E'])
 
         trade = AccountTrades(account=ta,
